Run everything in the browser
Irmin can run in the browser using js_of_ocaml, so it's simpler to do
everything there. We can add the server back in later, once everything
else is working.

Switched from ocaml-cow to TyXML, as it integrates better with
js_of_ocaml (and also ensures the output is valid HTML 5).