Animate moves correctly
Need to do this with code because we can't know the correct max-height
values for the CSS keyframes.

When Slow_set detects a rename (move), it creates a ref cell and gives
it to both the old and new states. The animation code for removal adds
in the height, which is known by this point. Later, the fade-in code
reads it and so discovers how big the element new will be.