Added `GroupItem type to widget trees
This is for labels (like `Group) that are actually nodes.
Unlike groups, renaming a GroupItem renames in place rather than
removing the old group and adding a new one.
Unlike items, a GroupItem may occur in several places in the tree.

For example, in the Work tab if Design/MyProg/X is replaced by
Code/MyProg/X then we remove the old MyProg and add a new one, but we
don't consider it to be a move (and don't animate it as such).