Support for merging
Previously, we could only merge two files if they were identical, and
could often fail even then. Using the new Irmin support for custom
merges, we can now always generate a valid merge of any two branches.

Note: this also fixes bug in init code where we create a node linking to
a context before the context, but fix it by the time we merge. The new
code detects this.