Improve contact display
Removed annoying "A contact" / "A context" message in details panel.
Instead, include it in the footer ("Contact created ...", etc).

Group contacts' nodes as "Responsible for" and "Contact for".

Show only "Responsible for" nodes in the Contacts tree.

The add button in the tree now starts the action in the
`Waiting_for_contact state.