Freshen item revision before updating
Each item records the revision it came from. To avoid unnecessary
updates to the GUI, we don't consider an item to have changed when
nothing except its revision is different. We also don't consider a
node as changed when its children change.

However, this means that we often try to commit against an old revision.
e.g.

- Item A is created in revision 1.
- Item B is created in revision 2.
- Item A (from revision 1) is renamed.

In this case, the final change's parent is revision 1, and we must now
merge it with revision 2. To avoid these unnecessary merges, we now
check whether the current version of the node is the same and use
the current version if so.

This also fixes the problem where e.g.

- Item A has child B (revision 1).
- B is deleted (revision 2).
- We attempt to delete A (from revision 1).

Here, the delete is rejected because A had a child at the time of its
last update.

To test this, this commit also makes the history output stable by
sorting by date if not constrained by topology.