Modernise build slightly
Use ocaml/opam2:debian-10-opam image to get opam 2 and use that to
simplify the installation a little.

The lwt problem has now been fixed in opam-repository.

As the build now runs as "opam", just add the build command to the
Dockerfile in the Travis tests.

Note: we still use aspcud, to avoid solver timeouts.